mcp-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.40.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
typing-extensions==4.8.0
httpx>=0.25.0
